[1.0] move xls test to correct test file
